Socket done, Elements addable
