+++ src/elements/basic.rs
@@ -4,7 +4,7 @@
 
 use conrod;
 
-use elements::*;
+use elements::{*, action::*};
 
 use std::sync::mpsc::{Sender};
 
@@ -45,6 +45,7 @@ pub struct Label {
     color: conrod::Color,
     background: Background,
 
+    is_setup: bool,
     frame: Frame<i32>,
     global_center: Vec2<i32>,
 
@@ -61,6 +62,7 @@ impl Label {
             font_size,
             color: conrod::color::BLACK,
             background: Background::None,
+            is_setup: false,
             frame: Frame::new(),
             global_center: Vec2::zero(),
             label_ids: None,
@@ -71,7 +73,10 @@ impl Label {
 impl Element for Label {
     fn setup(&mut self, ui: &mut conrod::Ui) {
         self.label_ids = Some(LabelIds::new(ui.widget_id_generator()));
+        self.is_setup = true;
+        if DEBUG { println!("Label --- setup()"); }
     }
+    fn is_setup(&self) -> bool { self.is_setup }
 
     fn build_window(&self, ui: &mut conrod::UiCell) {
         use conrod::{widget, Positionable, Colorable, Widget};
@@ -177,6 +182,7 @@ pub struct Button {
 
     //receive_fn: Box<Fn(&mut Element, ActionMsg)>,
 
+    is_setup: bool,
     global_center: Vec2<i32>,
     frame: Frame<i32>,
 
@@ -196,6 +202,7 @@ impl Button {
             senders: Vec::new(),
             //receive_fn: rfun,
 
+            is_setup: false,
             global_center: Vec2::zero(),
             frame: Frame::new(),
             button_ids: None,
@@ -263,7 +270,9 @@ impl ActionSendable for Button {
 impl Element for Button {
     fn setup(&mut self, ui: &mut conrod::Ui) {
         self.button_ids = Some(ButtonIds::new(ui.widget_id_generator()));
+        self.is_setup = true;
     }
+    fn is_setup(&self) -> bool { self.is_setup }
 
     fn build_window(&self, ui: &mut conrod::UiCell) {
         use conrod::{widget, Positionable, Colorable, Widget, Sizeable, Labelable, Borderable};
