--- src/elements/socket.rs
@@ -1,14 +1,0 @@
-
-
-use conrod;
-
-use elements::*;
-
-
-
-const DEBUG: bool = false;
-
-
-
-
-
